Added transaction isolation for Review deletion to be Serialisable. This ensures that an administrator experiences the correct behaviour on deleting a review

+++ src/java/social_network/ReviewDBAO.java
@@ -149,7 +149,6 @@ public class ReviewDBAO {
             throws SQLException, ClassNotFoundException {
         Connection con = null;
         PreparedStatement pStmt = null;
-        ResultSet resultSet = null;
         try {
             con = getConnection();
             pStmt = con.prepareStatement("INSERT INTO Reviews VALUES (?, ?, ?, ?, ?)");
@@ -175,25 +174,45 @@ public class ReviewDBAO {
  
    }
     
-     public static void deleteReview(String patient_login, String doctor_login, String datetime) 
+     public static boolean deleteReview(String patient_login, String doctor_login, String datetime) 
             throws SQLException, ClassNotFoundException {
         Connection con = null;
         PreparedStatement pStmt = null;
         ResultSet resultSet = null;
+        boolean isDeleted = false;
         try {
             con = getConnection();
-            pStmt = con.prepareStatement("DELETE FROM Reviews WHERE patient_login=? AND doctor_login=? AND datetime=?");
+            pStmt = con.prepareStatement("SELECT * FROM review_view WHERE patient_login=? and doctor_login=? AND datetime=?");
             pStmt.setString(1, patient_login);
             pStmt.setString(2, doctor_login);
             pStmt.setString(3, datetime);
-            pStmt.execute();
-        }finally {
+            resultSet = pStmt.executeQuery();
+            if(resultSet.first()) {
+                con.setAutoCommit(false);
+                con.setTransactionIsolation(Connection.TRANSACTION_SERIALIZABLE);
+                pStmt = con.prepareStatement("DELETE FROM Reviews WHERE patient_login=? AND doctor_login=? AND datetime=?");
+                pStmt.setString(1, patient_login);
+                pStmt.setString(2, doctor_login);
+                pStmt.setString(3, datetime);
+                pStmt.executeUpdate();
+                con.commit();
+                con.setAutoCommit(true);
+                isDeleted = true;
+            }else {
+                isDeleted = false;
+            }
+        }catch(SQLException se) {
+            System.out.print(se.toString());
+            con.rollback();            
+        } finally {
+            //con.setAutoCommit(true);
             if (pStmt != null) {
                 pStmt.close();
             }
             if (con != null) {
                 con.close();
             }    
+            return isDeleted;
         }
     }
